# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Alert API - Government Jobs Aggregator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  department: string;
  category: string;
  states: string[];
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  ageRelaxation?: string | null;
  /** @nullable */
  qualification?: string | null;
  /** @nullable */
  applicationStartDate?: string | null;
  /** @nullable */
  lastDate?: string | null;
  /** @nullable */
  examDate?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  notificationPdfUrl?: string | null;
  /** @nullable */
  sourceName?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  isVerified: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  salaryRange?: string | null;
  /** @nullable */
  vacancies?: number | null;
  /** @nullable */
  photoRequirements?: string | null;
  /** @nullable */
  signatureRequirements?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface JobCount {
  count: number;
}

export interface Category {
  slug: string;
  name: string;
  count: number;
  color: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetJobsParams = {
  /**
   * Search term for title or department
   */
  search?: string;
  /**
   * Filter by category slug
   */
  category?: string;
  /**
   * Filter by state name
   */
  state?: string;
  /**
   * Filter by qualification
   */
  qualification?: string;
  page?: number;
  limit?: number;
};