# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Job Alert API - Government Jobs Aggregator
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List jobs
      description: Get verified, non-expired job listings with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search term for title or department
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state name
        - name: qualification
          in: query
          schema:
            type: string
          description: Filter by qualification
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
...
  /jobs/{id}:
    get:
      operationId: getJobById
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        category:
          type: string
        states:
          type: array
          items:
            type: string
        ageMin:
          type: ["integer", "null"]
        ageMax:
          type: ["integer", "null"]
        ageRelaxation:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        applicationStartDate:
          type: ["string", "null"]
        lastDate:
          type: ["string", "null"]
        examDate:
          type: ["string", "null"]
        applyUrl:
          type: ["string", "null"]
        notificationPdfUrl:
          type: ["string", "null"]
        sourceName:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        description:
          type: ["string", "null"]
        salaryRange:
          type: ["string", "null"]
        vacancies:
          type: ["integer", "null"]
        photoRequirements:
          type: ["string", "null"]
        signatureRequirements:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - department
        - category
        - states
        - isVerified
        - createdAt
        - updatedAt
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - jobs
        - total
        - page
        - limit
        - totalPages
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error