# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Alert API - Government Jobs Aggregator
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  title: string;
  department: string;
  category: string;
  states: string[];
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  ageRelaxation?: string | null;
  /** @nullable */
  qualification?: string | null;
  /** @nullable */
  applicationStartDate?: string | null;
  /** @nullable */
  lastDate?: string | null;
  /** @nullable */
  examDate?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  notificationPdfUrl?: string | null;
  /** @nullable */
  sourceName?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  isVerified: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  salaryRange?: string | null;
  /** @nullable */
  vacancies?: number | null;
  /** @nullable */
  photoRequirements?: string | null;
  /** @nullable */
  signatureRequirements?: string | null;
  createdAt: string;
  updatedAt: string;
}