# lib/api-zod/src/generated/types/jobListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Alert API - Government Jobs Aggregator
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}